feat: impl a mini mosec to work with model in subprocess
